Update name image

+++ lib/apis/apis.dart
@@ -420,7 +420,9 @@ class APIs {
         .toList();
     List<MessageChat> listSearchMessage = [];
     for (MessageChat message in listMessage) {
-      if (TiengViet.parse(message.msg!).toLowerCase().contains(TiengViet.parse(_enteredWord).toLowerCase())) {
+      if (TiengViet.parse(message.msg!)
+          .toLowerCase()
+          .contains(TiengViet.parse(_enteredWord).toLowerCase())) {
         listSearchMessage.add(message);
       }
     }
@@ -606,10 +608,15 @@ class APIs {
         .snapshots();
   }
 
-  static Future<void> updateImageUser(String userID, File imageFile) async {
+  static Future<void> updateImageUser(
+      String userID, File imageFile, String imgID) async {
     final storageRef =
         await fireStorage.ref().child('user_images').child('${userID}.jpg');
-    await storageRef.delete();
+
+    if (imgID != '') {
+      await storageRef.delete();
+    }
+
     String imgURL = await saveMedia(0, userID, imageFile, 'user_images');
     return await firestore
         .collection('user')
@@ -618,12 +625,14 @@ class APIs {
   }
 
   static Future<void> updateImageChatRoom(
-      String chatRoomID, File imageFile) async {
+      String chatRoomID, File imageFile, String imgID) async {
     final storageRef = await fireStorage
         .ref()
         .child('chatroom_images')
         .child('${chatRoomID}.jpg');
-    await storageRef.delete();
+    if (imgID != '') {
+      await storageRef.delete();
+    }
     String imgURL =
         await saveMedia(0, chatRoomID, imageFile, 'chatroom_images');
     return await firestore
